chore: organize project structure and file paths
