+++ frontend/src/App.tsx
@@ -1,4 +1,4 @@
-import './App.css';
+import './styles/App.css';
 import { BrowserRouter } from 'react-router-dom';
 import { AppRoutes } from './components/appRoutes';
