+++ frontend/src/main.tsx
@@ -1,7 +1,7 @@
 import { createRoot } from 'react-dom/client'
 
 import App from './App'
-import './main.css'
+import './styles/main.css'
 
 createRoot(document.getElementById('root')!).render(
   <App />
